requirements.txt
pymarkdownlnt==0.9.12
